Fix visibility after restructure
